feat: 리뷰 의견 반영 - repository 구현체 Domain 이동 및 stub respositroy 이동
