+++ Project/Projects/Feature/Calendar/SampleApp/Sources/StubCalendarRepository.swift
@@ -9,6 +9,7 @@
 import Foundation
 import ArchiveFoundation
 import UIComponents
+import Domain
 
 public final class StubCalendarRepositoryImpl: CalendarRepository {
